Update for new database structure
See pihole doc for more info

+++ clean.sh
@@ -1,12 +1,20 @@
-#!/bin/sh
-
+#!/bin/bash
 
 #	This is a very simple script that removes some querries from the pihole querry every time this was executed.
 #	This removes them in both the long term querry log, the 24-hour log and the recent 100 querries.
 
-# Prints the date every echo. Add a "#" before the following line to remove it.
-alias echo='echo $(date)'
-echo Script Started, both code should be a 0. If not there is a problem.
+# Check user root
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root. (Did you forgot sudo?)"
+  exit 1
+fi
+
+rmQuery () {
+        sqlite3 /etc/pihole/pihole-FTL.db "delete from query_storage where domain in (select id from domain_by_id where domain like '$1');"
+}
+
+echo Script Started at $(date)
+echo 'Both code should be a 0. If not there is a problem.'
 # Stops FTL, So that we are not writing to a live database.
 systemctl stop pihole-FTL
 echo stop code = $?
@@ -14,11 +22,14 @@ cd /etc/pihole
 # Starts removing querries.
 
 # Remove whatever ends with "in-addr.arpa"
-sqlite3 pihole-FTL.db "delete from query_storage where domain like '%in-addr.arpa';"
+rmQuery '%in-addr.arpa'
+
 # Remove whatever contains "ip6.arpa"
-sqlite3 pihole-FTL.db "delete from query_storage where domain like '%ip6.arpa%';"
+rmQuery '%ip6.arpa%'
+
 # Remove "empty"
-sqlite3 pihole-FTL.db "delete from query_storage where domain like 'empty';"
+rmQuery 'empty'
+
 echo [✓] Part 1 ok.
 # You may add Part 2 here.
 
@@ -26,4 +37,4 @@ echo [✓] Part 1 ok.
 systemctl restart pihole-FTL
 echo restart code = $?
 echo [✓] Done.
-echo Script Ended
+echo Script Ended at $(date)
